Autotools build infrastructure for kinetic-protocol libraries
Provide the build support to build kinetic-protocol specific libraries for
utilization in C, C++ & python applications.

The autotools also enables the build setup in cross and various
architectures.

The build supports building shared libraries for C, C++ and python.
The individual libraries can be enabled or disabled through the
configure options.

The Autotools build procedure is documented in the README.build
file. It provides information on build steps from scratch for
a native build.

Signed-off-by: Kinjal Patel <kinjal.patel@taec.toshiba.com>
Signed-off-by: Nitin A Kamble <Nitin.Kamble@taec.toshiba.com>